new AccountCustomer created V-101

+++ accountservice/src/main/java/com/acheter/accountservice/repository/SystemUserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.acheter.accountservice.entities.System_user;
 
-public interface SystemUserRepository extends JpaRepository<System_user, Integer>{
+public interface SystemUserRepository extends JpaRepository<System_user, Long>{
 
 	
 }
